backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Deprecated==1.2.14
Django==4.2.7
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jwcrypto==1.5.0
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
